requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.10.0
pyjwt==2.8.0
httpx==0.26.0
celery==5.3.6
redis==5.0.1
python-socketio==5.11.0
